+++ R/utils.R
@@ -91,3 +91,40 @@ cli_escape <- function(x) {
   x <- gsub("}", "}}", x, fixed = TRUE)
   x
 }
+
+check_labels <- function(
+  labels,
+  arg = caller_arg(labels),
+  call = caller_env()
+) {
+  if (is.null(labels)) {
+    return(invisible())
+  }
+
+  if (!is.list(labels) || any(names2(labels) == "")) {
+    stop_input_type(
+      labels,
+      "a named list",
+      allow_null = TRUE,
+      arg = arg,
+      call = call
+    )
+  }
+  is_string <- vapply(labels, is_string, logical(1))
+  if (!all(is_string)) {
+    stop_input_type(
+      labels,
+      "a named list of strings",
+      arg = arg,
+      call = call
+    )
+  }
+
+  invisible()
+}
+
+cli_escape <- function(x) {
+  x <- gsub("{", "{{", x, fixed = TRUE)
+  x <- gsub("}", "}}", x, fixed = TRUE)
+  x
+}